// src/app/models/product.ts
export interface Product {
    _id: string
    sku: string
    attribute_set_code: string
    categories: string
    name: string
    description: string
    weight: number
    price: number
    meta_title: string
    meta_description: string
    gtin: number
    brand: string
    product_brand: string
    delivery_information: string
    exclude_from_google_feed: string
    finish: string
    google_custom_label_0: string
    google_custom_label_1: string
    google_title: string
    google_product_category: string
    product_type_0: string
    product_type_1: string
    size: string
    backset_size: string
    fixing_centre: string
    style: string
    unit: string
    related_skus: string
    related_position: string
    upsell_skus: string
    upsell_position: string
  }
  